Moved ConsignmentData from form to model
