+++ app/controllers/CreateConsignmentController.scala
@@ -1,11 +1,9 @@
 package controllers
 
-import akka.http.scaladsl.model.headers.RawHeader
-import forms.ConsignmentForm
 import graphql.GraphQLClientProvider
 import graphql.codegen.CreateConsignment.createConsignment._
-
 import javax.inject.{Inject, _}
+import model.CreateConsignmentData
 import play.api.Configuration
 import play.api.mvc._
 
@@ -19,13 +17,13 @@ class CreateConsignmentController @Inject()(
                                              implicit val ex: ExecutionContext) extends AbstractController(controllerComponents) with play.api.i18n.I18nSupport {
 
   def index(seriesId:Int) = Action.async { implicit request: Request[AnyContent] =>
-    Future(Ok(views.html.createConsignments(ConsignmentForm.form, seriesId)))
+    Future(Ok(views.html.createConsignments(CreateConsignmentData.form, seriesId)))
   }
 
   def submit() = Action.async { implicit request =>
 
     val appSyncClient = client.graphqlClient(List())
-    val form = ConsignmentForm.form.bindFromRequest
+    val form = CreateConsignmentData.form.bindFromRequest
 
     val vars = Variables(form.get.consignmentName,form.get.seriesId,request.session.get("username").get,form.get.transferringBody)
     appSyncClient.query[Data,Variables](document, vars).result.map(result => result match {
