+++ app/model/CreateConsignmentData.scala
@@ -1,17 +1,17 @@
-package forms
+package model
 
 import play.api.data.Form
 import play.api.data.Forms._
 
-case class ConsignmentData(consignmentName: String, transferringBody: String, seriesId: Int)
+case class CreateConsignmentData(consignmentName: String, transferringBody: String, seriesId: Int)
 
-object ConsignmentData {
-  val form: Form[ConsignmentData] = Form(
+object CreateConsignmentData {
+  val form: Form[CreateConsignmentData] = Form(
     mapping(
       "consignmentName" -> text,
       "transferringBody" -> text,
       "seriesId" -> number
-    )(ConsignmentData.apply)(ConsignmentData.unapply)
+    )(CreateConsignmentData.apply)(CreateConsignmentData.unapply)
   )
 }
